refactor: add a new package
