+++ src/test/java/nextstep/subway/application/strategy/addition/AddSectionAfterUpStationStrategyTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.application.strategy;
+package nextstep.subway.application.strategy.addition;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -8,6 +8,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
+import nextstep.subway.application.strategy.addition.AddSectionAfterUpStationStrategy;
 import nextstep.subway.domain.model.Line;
 import nextstep.subway.domain.model.Section;
 import nextstep.subway.domain.model.Station;
